requirements.txt
appdirs==1.4.4
aspy.yaml==1.3.0
attrs==20.3.0
beautifulsoup4==4.8.2
blis==0.4.1
catalogue==1.0.0
certifi==2020.12.5
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
cymem==2.0.3
dataclasses==0.6
distlib==0.3.1
en-core-web-sm==2.2.5
filelock==3.0.12
future==0.18.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==1.5.10
idna==2.8
importlib-metadata==3.1.1
isodate==0.6.0
joblib==0.17.0
langdetect==1.0.7
lxml==4.6.2
more-itertools==8.1.0
murmurhash==1.0.5
nltk==3.4.5
nodeenv==1.5.0
numpy==1.18.1
Owlready2==0.24
packaging==20.7
Pillow==8.0.1
plac==1.1.3
pluggy==0.13.1
pre-commit==2.9.3
preshed==3.0.5
protobuf==3.14.0
py==1.9.0
pyparsing==2.4.7
pytest==5.3.5
PyYAML==5.3
rdflib==5.0.0
regex==2020.11.13
requests==2.22.0
rfc3986==1.4.0
sacremoses==0.0.43
sentencepiece==0.1.91
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
spacy==2.2.3
spacy-langdetect==0.1.2
spelchek==0.54
srsly==1.0.2
thinc==7.3.1
tokenizers==0.9.3
toml==0.10.2
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
tox==3.20.1
tqdm==4.54.1
transformers==3.5.1
translate==3.5.0
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.2.2
wasabi==0.8.0
wcwidth==0.2.5
zipp==3.4.0
